[dependencies] Fix the 404 'asyncio-dev' not found.
On this commit, I use @pokurt's fork of Pyrogram on 'asyncio-dev' branch.
To improve performance on running userbots in Pyrogram, I added 'TgCrypto'
as an dependency in this userbot.

While using TgCrypto in Pyrogram is optional as per the documentation,
it's recommended to install that package to improve performance
on crypto stuff.

Signed-off-by: Andrei Jiroh Eugenio Halili <andreijiroh@madebythepins.tk>